#802 rename calendar module to timeline
