+++ docs/str_funcs.md
@@ -8,7 +8,7 @@
 - a04_reason_logic: _active, _chore, _status, fcontext, fnigh, fopen, fstate, labor_title, pdivisor, pnigh, popen, pstate, rconcept_active_requisite, rcontext
 - a05_concept_logic: _active_hx, _all_acct_cred, _all_acct_debt, _awardheirs, _awardlines, _descendant_task_count, _factheirs, _fund_cease, _fund_onset, _fund_ratio, _gogo_calc, _healerlink_ratio, _is_expanded, _kids, _range_evaluated, _reasonheirs, _stop_calc, _uid, addin, begin, close, concept_label, concept_rope, denom, gogo_want, healer_name, healerlink, mass, morph, numor, problem_bool, stop_want, task
 - a06_plan_logic: _keeps_buildable, _keeps_justified, _offtrack_fund, _offtrack_kids_mass_set, _rational, _reason_rcontexts, _sum_healerlink_share, _tree_traverse_count, acct_pool, ancestors, attributes, credor_respect, debtor_respect, dimen, dimens, jkeys, last_pack_id, mandate, max_tree_traverse, plan_acct_membership, plan_acctunit, plan_concept_awardlink, plan_concept_factunit, plan_concept_healerlink, plan_concept_laborlink, plan_concept_reason_premiseunit, plan_concept_reasonunit, plan_conceptunit, planunit, tally, timeline
-- a07_calendar_logic: c100, c400_clean, c400_leap, c400_number, creg, cumulative_day, day, days, five, hour, hours_config, monthday_distortion, months_config, time, timeline_label, week, weekdays_config, weeks, year, yr1_jan1_offset, yr4_clean, yr4_leap
+- a07_timeline_logic: c100, c400_clean, c400_leap, c400_number, creg, cumulative_day, day, days, five, hour, hours_config, monthday_distortion, months_config, time, timeline_label, week, weekdays_config, weeks, year, yr1_jan1_offset, yr4_clean, yr4_leap
 - a08_plan_atom_logic: DELETE, atom_hx, class_type, column_order, crud, jvalues, nesting_order, normal_specs, normal_table_name
 - a09_pack_logic: event_int, face_name
 - a10_plan_calc: jmetrics, plan_groupunit
