--- src/a07_calendar_logic/test/_util/a07_env.py
@@ -1,2 +1,0 @@
-def get_module_examples_dir() -> str:
-    return "src/a07_calendar_logic/test/_util"
