+++ src/a07_timeline_logic/test/_util/a07_env.py
@@ -1,0 +1,2 @@
+def get_module_examples_dir() -> str:
+    return "src/a07_timeline_logic/test/_util"
