+++ src/a07_timeline_logic/test/_util/test_a07_str.py
@@ -1,4 +1,4 @@
-from src.a07_calendar_logic.test._util.a07_str import (
+from src.a07_timeline_logic.test._util.a07_str import (
     c100_str,
     c400_clean_str,
     c400_leap_str,
