+++ src/a07_timeline_logic/timeline.py
@@ -31,7 +31,7 @@ class C400Constants:
 
 
 def get_c400_constants() -> C400Constants:
-    c400_constants_path = create_path("src/a07_calendar_logic", "c400_constants.json")
+    c400_constants_path = create_path("src/a07_timeline_logic", "c400_constants.json")
     c400_dict = open_json(c400_constants_path)
     return C400Constants(
         day_length=c400_dict.get("day_length"),
@@ -538,7 +538,7 @@ def calendarstr_shop(x_planunit: PlanUnit, time_range_root_rope: str, x_min: int
 
 def config_file_dir() -> str:
     src_dir = create_path(os_getcwd(), "src")
-    return create_path(src_dir, "a07_calendar_logic")
+    return create_path(src_dir, "a07_timeline_logic")
 
 
 def get_default_timeline_config_filename() -> str:
