+++ src/a17_idea_logic/test/_util/idea_df_examples.py
@@ -13,7 +13,7 @@ from src.a02_finance_logic.test._util.a02_str import (
     vow_label_str,
 )
 from src.a03_group_logic.test._util.a03_str import respect_bit_str
-from src.a07_calendar_logic.test._util.a07_str import (
+from src.a07_timeline_logic.test._util.a07_str import (
     c400_number_str,
     monthday_distortion_str,
     timeline_label_str,
