+++ src/a17_idea_logic/test/test_z_vow_build/test_idea_vow_build.py
@@ -1,6 +1,6 @@
 from src.a00_data_toolbox.file_toolbox import create_path
-from src.a07_calendar_logic.test._util.calendar_examples import get_five_config
-from src.a07_calendar_logic.timeline import (
+from src.a07_timeline_logic.test._util.calendar_examples import get_five_config
+from src.a07_timeline_logic.timeline import (
     get_default_timeline_config_dict,
     timelineunit_shop,
 )
